mfa/pyproject.toml
[project]
name = "mfa-agents"
version = "0.1.0"
description = "Microsoft Agent Framework examples - translated from Google ADK"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-framework>=0.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "aiohttp>=3.9.0",
    "agent-framework-devui>=1.0.0b251204",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
